+++ src/torchmetrics/functional/text/_deprecated.py
@@ -24,12 +24,12 @@ from torchmetrics.functional.text.ter import translation_edit_rate
 from torchmetrics.functional.text.wer import word_error_rate
 from torchmetrics.functional.text.wil import word_information_lost
 from torchmetrics.functional.text.wip import word_information_preserved
-from torchmetrics.utilities.imports import _TRANSFORMERS_GREATER_EQUAL_4_4
+from torchmetrics.utilities.imports import _TRANSFORMERS_AVAILABLE
 from torchmetrics.utilities.prints import _deprecated_root_import_func
 
 __doctest_requires__ = {("_rouge_score"): ["nltk"]}
 
-if not _TRANSFORMERS_GREATER_EQUAL_4_4:
+if not _TRANSFORMERS_AVAILABLE:
     __doctest_skip__ = ["_bert_score", "_infolm"]
 
 SQUAD_SINGLE_TARGET_TYPE = dict[str, Union[str, dict[str, Union[list[str], list[int]]]]]