+++ src/torchmetrics/text/bert.py
@@ -28,7 +28,7 @@ from torchmetrics.metric import Metric
 from torchmetrics.utilities import rank_zero_warn
 from torchmetrics.utilities.checks import _SKIP_SLOW_DOCTEST, _try_proceed_with_timeout
 from torchmetrics.utilities.data import dim_zero_cat
-from torchmetrics.utilities.imports import _MATPLOTLIB_AVAILABLE, _TRANSFORMERS_GREATER_EQUAL_4_4
+from torchmetrics.utilities.imports import _MATPLOTLIB_AVAILABLE, _TRANSFORMERS_AVAILABLE
 from torchmetrics.utilities.plot import _AX_TYPE, _PLOT_OUT_TYPE
 
 if not _MATPLOTLIB_AVAILABLE:
@@ -37,7 +37,7 @@ if not _MATPLOTLIB_AVAILABLE:
 # Default model recommended in the original implementation.
 _DEFAULT_MODEL: str = "roberta-large"
 
-if _SKIP_SLOW_DOCTEST and _TRANSFORMERS_GREATER_EQUAL_4_4:
+if _SKIP_SLOW_DOCTEST and _TRANSFORMERS_AVAILABLE:
     from transformers import AutoModel, AutoTokenizer
 
     def _download_model_for_bert_score() -> None:
@@ -201,10 +201,10 @@ class BERTScore(Metric):
             self.tokenizer = user_tokenizer
             self.user_tokenizer = True
         else:
-            if not _TRANSFORMERS_GREATER_EQUAL_4_4:
+            if not _TRANSFORMERS_AVAILABLE:
                 raise ModuleNotFoundError(
                     "`BERTScore` metric with default tokenizers requires `transformers` package be installed."
-                    " Either install with `pip install transformers>=4.4` or `pip install torchmetrics[text]`."
+                    " Either install with `pip install transformers` or `pip install torchmetrics[text]`."
                 )
             from transformers import AutoTokenizer