+++ tests/unittests/multimodal/test_clip_iqa.py
@@ -25,7 +25,7 @@ from torchvision.transforms import PILToTensor
 
 from torchmetrics.functional.multimodal.clip_iqa import clip_image_quality_assessment
 from torchmetrics.multimodal.clip_iqa import CLIPImageQualityAssessment
-from torchmetrics.utilities.imports import _PIQ_GREATER_EQUAL_0_8, _TRANSFORMERS_GREATER_EQUAL_4_10
+from torchmetrics.utilities.imports import _PIQ_GREATER_EQUAL_0_8, _TRANSFORMERS_AVAILABLE
 from unittests._helpers import skip_on_connection_issues
 from unittests._helpers.testers import MetricTester
 from unittests.image import _SAMPLE_IMAGE
@@ -72,7 +72,7 @@ def _reference_clip_iqa(preds, target, reduce=False):
 
 
 @pytest.mark.skipif(not _PIQ_GREATER_EQUAL_0_8, reason="metric requires piq>=0.8")
-@pytest.mark.skipif(not _TRANSFORMERS_GREATER_EQUAL_4_10, reason="test requires transformers>=4.10")
+@pytest.mark.skipif(not _TRANSFORMERS_AVAILABLE, reason="test requires transformers")
 class TestCLIPIQA(MetricTester):
     """Test clip iqa metric."""
 
@@ -105,7 +105,7 @@ class TestCLIPIQA(MetricTester):
 
 @skip_on_connection_issues()
 @pytest.mark.skipif(not _PIQ_GREATER_EQUAL_0_8, reason="metric requires piq>=0.8")
-@pytest.mark.skipif(not _TRANSFORMERS_GREATER_EQUAL_4_10, reason="test requires transformers>=4.10")
+@pytest.mark.skipif(not _TRANSFORMERS_AVAILABLE, reason="test requires transformers")
 @pytest.mark.skipif(not os.path.isfile(_SAMPLE_IMAGE), reason="test image not found")
 def test_for_correctness_sample_images():
     """Compare the output of the function with the output of the reference implementation."""
@@ -122,7 +122,7 @@ def test_for_correctness_sample_images():
 
 @skip_on_connection_issues()
 @pytest.mark.skipif(not _PIQ_GREATER_EQUAL_0_8, reason="metric requires piq>=0.8")
-@pytest.mark.skipif(not _TRANSFORMERS_GREATER_EQUAL_4_10, reason="test requires transformers>=4.10")
+@pytest.mark.skipif(not _TRANSFORMERS_AVAILABLE, reason="test requires transformers")
 @pytest.mark.parametrize(
     "model",
     [
@@ -149,7 +149,7 @@ def test_other_models(model):
 
 @skip_on_connection_issues()
 @pytest.mark.skipif(not _PIQ_GREATER_EQUAL_0_8, reason="metric requires piq>=0.8")
-@pytest.mark.skipif(not _TRANSFORMERS_GREATER_EQUAL_4_10, reason="test requires transformers>=4.10")
+@pytest.mark.skipif(not _TRANSFORMERS_AVAILABLE, reason="test requires transformers")
 @pytest.mark.parametrize(
     "prompts",
     [
@@ -201,7 +201,7 @@ def test_prompt(prompts):
 
 @skip_on_connection_issues()
 @pytest.mark.skipif(not _PIQ_GREATER_EQUAL_0_8, reason="metric requires piq>=0.8")
-@pytest.mark.skipif(not _TRANSFORMERS_GREATER_EQUAL_4_10, reason="test requires transformers>=4.10")
+@pytest.mark.skipif(not _TRANSFORMERS_AVAILABLE, reason="test requires transformers")
 def test_plot_method():
     """Test the plot method of CLIPScore separately in this file due to the skipping conditions."""
     metric = CLIPImageQualityAssessment()